roles y usuarios de prueba

+++ demo/src/main/java/com/example/demo/model/Valoracion.java
@@ -19,7 +19,7 @@ public class Valoracion {
 
     private String comentario;
 
-    private Long calificacion;
+    private Integer calificacion;
 
     @JsonIgnore
     @DBRef
